Improve pointer propagation and class layout recovery

+++ decomp/ST.exe/functions/0041D6C0/decomp.c
@@ -14,7 +14,7 @@ int __thiscall FUN_0041d6c0(void *this,uint param_1)
   }
   iVar1 = (int)*(short *)((int)this + 0x5b);
   iVar2 = (int)*(short *)((int)this + 0x5d);
-  if ((((-1 < iVar1) && (iVar1 < (int)g_visibleClass_00802A88->field_0020)) && (-1 < iVar2)) &&
+  if ((((-1 < iVar1) && (iVar1 < g_visibleClass_00802A88->field_0020)) && (-1 < iVar2)) &&
      (((iVar2 < g_visibleClass_00802A88->field_0024 && (param_1 < 8)) &&
       (*(int *)(&g_visibleClass_00802A88->field_0xd4 + param_1 * 4) != 0)))) {
     return (uint)(*(char *)(g_visibleClass_00802A88->field_0020 * iVar2 +
