+++ decomp/ST.exe/functions/004D8F80/decomp.c
@@ -5,7 +5,7 @@ undefined4 __fastcall FUN_004d8f80(AnonShape_004D8F80_1255A4D6 *param_1)
   int iVar1;
 
   if ((int)param_1->field_04EC < 100) {
-    if (param_1->field_04D8 + 0x4b <= (uint)PTR_00802a38->field_00E4) {
+    if (param_1->field_04D8 + 0x4b <= PTR_00802a38->field_00E4) {
       iVar1 = param_1->field_04EC + 1;
       param_1->field_04D8 = PTR_00802a38->field_00E4;
       param_1->field_04EC = iVar1;
