+++ decomp/ST.exe/functions/004E4B10/decomp.c
@@ -3,7 +3,7 @@ undefined4 __fastcall FUN_004e4b10(AnonShape_004E4B10_C52B4B94 *param_1)
 
 {
   if ((param_1->field_0245 == 0) && (param_1->field_04D4 < 0x28)) {
-    if (param_1->field_04D8 + 0x1c <= (uint)PTR_00802a38->field_00E4) {
+    if (param_1->field_04D8 + 0x1c <= PTR_00802a38->field_00E4) {
       param_1->field_04D8 = PTR_00802a38->field_00E4;
       param_1->field_04D4 = param_1->field_04D4 + 1;
     }
