+++ decomp/ST.exe/functions/00579350/decomp.c
@@ -113,7 +113,7 @@ LAB_00579591:
     param_1->field_023D = 0;
   }
 LAB_005795a9:
-  if (2 < (uint)(PTR_00802a38->field_00E4 - param_1->field_0235)) {
+  if (2 < PTR_00802a38->field_00E4 - param_1->field_0235) {
     param_1->field_0235 = PTR_00802a38->field_00E4;
     if ((*(short *)&param_1->field_0x6c == 7) || (*(short *)&param_1->field_0x6c < 3)) {
       param_1->field_0231 = (param_1->field_0231 + 1) % 6;
