+++ decomp/ST.exe/functions/00621580/decomp.c
@@ -14,7 +14,7 @@ void __thiscall FUN_00621580(void *this,uint param_1,int param_2)
     uVar1 = thunk_FUN_006216e0(this,param_2);
   }
   if ((local_8 != 0) && (-1 < (int)uVar1)) {
-    if (10 < (uint)(PTR_00802a38->field_00E4 - *(int *)((int)this + 0x106))) {
+    if (10 < PTR_00802a38->field_00E4 - *(int *)((int)this + 0x106)) {
       thunk_FUN_00621780(this,uVar1);
       return;
     }
